bsg: add bsg heler library
Patchwork-id: 38791
O-Subject: [RHEL6.2 PATCH 14/20] bsg: add bsg heler library
Bugzilla: 634584
RH-Acked-by: Rob Evers <revers@redhat.com>

From: Mike Christie <mchristi@redhat.com>

This moves the FC classes bsg code to the block layer and
makes it a lib so that other classes like iscsi and SAS can use it.

It is helpful because working with the request queue, bios,
creating scatterlists, etc are a pain that the LLD does not
have to worry about with normal IOs and should not have to
worry about for bsg requests.

Jens has said is going to merge the patch:
http://marc.info/?l=linux-scsi&m=131153238516003&w=2

This version of the patch is slightly different from the
upstream version due to the upstream patch adding some fields
to the request_queue.