Introduce hierarchical task accounting
Patchwork-id: 38806
O-Subject: [RHEL6.2 PATCH 02/23] Introduce hierarchical task accounting
Bugzilla: 632509
RH-Acked-by: Rik van Riel <riel@redhat.com>

Introduce hierarchical task accounting for the group scheduling case in CFS, as
well as promoting the responsibility for maintaining rq->nr_running to the
scheduling classes.

The primary motivation for this is that with scheduling classes supporting
bandwidth throttling it is possible for entities participating in throttled
sub-trees to not have root visible changes in rq->nr_running across activate
and de-activate operations.  This in turn leads to incorrect idle and
weight-per-task load balance decisions.

This also allows us to make a small fixlet to the fastpath in pick_next_task()
under group scheduling.

Note: this issue also exists with the existing sched_rt throttling mechanism.
This patch does not address that.

Signed-off-by: Paul Turner <pjt@google.com>
Reviewed-by: Hidetoshi Seto <seto.hidetoshi@jp.fujitsu.com>
Signed-off-by: Larry Woodman <lwoodman@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>