Throttled tasks are invisisble to cpu-offline
Patchwork-id: 38813
O-Subject: [RHEL6.2 PATCH 13/23] Throttled tasks are invisisble to cpu-offline
Bugzilla: 632509
RH-Acked-by: Rik van Riel <riel@redhat.com>

Throttled tasks are invisisble to cpu-offline since they are not eligible for
selection by pick_next_task().  The regular 'escape' path for a thread that is
blocked at offline is via ttwu->select_task_rq, however this will not handle a
throttled group since there are no individual thread wakeups on an unthrottle.

Resolve this by unthrottling offline cpus so that threads can be migrated.

Signed-off-by: Paul Turner <pjt@google.com>
Reviewed-by: Hidetoshi Seto <seto.hidetoshi@jp.fujitsu.com>
Signed-off-by: Larry Woodman <lwoodman@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>