script: If type not given fields apply to all event types
Patchwork-id: 38964
O-Subject: [PATCH RHEL6.2 BZ725522 05/67] perf script: If type not given fields
	apply to all event types
Bugzilla: 725522
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 725522
https://bugzilla.redhat.com/show_bug.cgi?id=725522

upstream status
---------------
commit 2c9e45f7a287384e1382932597e41a9a567811ba
Author: David Ahern <daahern@cisco.com>
Date:   Thu Mar 17 10:03:21 2011 -0600

description
-----------
Allow:
  perf script -f <fields>

to be equivalent to:
  perf script -f trace:<fields> -f sw:<fields> -f hw:<fields>

i.e., the specified fields apply to all event types if the type string
is not given.

The field (-f) arguments are processed in the order received. A later
usage can reset a prior request. e.g.,

  -f trace: -f comm,tid,time,sym

The first -f suppresses trace events (field list is ""), but then the second
invocation sets the fields to comm,tid,time,sym. In this case a warning is
given to the user:

  "Overriding previous field request for all events."

Alternativey, consider the order:

  -f comm,tid,time,sym -f trace:

The first -f sets the fields for all events and the second -f suppresses trace
events. The user is given a warning message about the override, and the result
of the above is that only S/W and H/W events are displayed with the given
fields.

For the 'wildcard' option if a user selected field is invalid for an event
type, a message is displayed to the user that the option is ignored for that
type. For example:

  perf script -f comm,tid,trace 2>&1 | less
  'trace' not valid for hardware events. Ignoring.
  'trace' not valid for software events. Ignoring.

Alternatively, if the type is given an invalid field is specified it is an
error. For example:

    perf script -v -f sw:comm,tid,trace 2>&1 | less
    'trace' not valid for software events.

At this point usage is displayed, and perf-script exits.

Finally, a user may not set fields to none for all event types.
i.e., -f "" is not allowed.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>