tools: Makefile: PYTHON{, _CONFIG} to bandage Python 3 incompatibility
Patchwork-id: 38924
O-Subject: [PATCH RHEL6.2 BZ725522 09/67] perf tools: Makefile: PYTHON{,
	_CONFIG} to bandage Python 3 incompatibility
Bugzilla: 725522
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 725522
https://bugzilla.redhat.com/show_bug.cgi?id=725522

upstream status
---------------
commit ced465c400b23656ef2c4fbfb4add0e5b92e3d97
Author: Michael Witten <mfwitten@gmail.com>
Date:   Sat Apr 2 21:46:09 2011 +0000

description
-----------
Currently, Python 3 is not supported by perf's code; this
can cause the build to fail for systems that have Python 3
installed as the default python:

  python{,-config}

The Correct Solution is to write compatibility code so that
Python 3 works out-of-the-box.

However, users often have an ancillary Python 2 installed:

  python2{,-config}

Therefore, a quick fix is to allow the user to specify those
ancillary paths as the python binaries that Makefile should
use, thereby avoiding Python 3 altogether; as an added benefit,
the Python binaries may be installed in non-standard locations
without the need for updating any PATH variable.

This commit adds the ability to set PYTHON and/or PYTHON_CONFIG
either as environment variables or as make variables on the
command line; the paths may be relative, and usually only PYTHON
is necessary in order for PYTHON_CONFIG to be defined implicitly.
Some rudimentary error checking is performed when the user
explicitly specifies a value for any of these variables.

In addition, this commit introduces significantly robust makefile
infrastructure for working with paths and communicating with the
shell; it's currently only used for handling Python, but I hope
it will prove useful in refactoring the makefiles.

Thanks to:

  Raghavendra D Prabhu <rprabhu@wnohang.net>

for motivating this patch.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>