stat: Fail more clearly when an invalid modifier is specified
Patchwork-id: 38926
O-Subject: [PATCH RHEL6.2 BZ725522 18/67] perf stat: Fail more clearly when an
	invalid modifier is specified
Bugzilla: 725522
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 725522
https://bugzilla.redhat.com/show_bug.cgi?id=725522

upstream status
---------------
commit ceb53fbf6dbb1df26d38379a262c6981fe73dd36
Author: Ingo Molnar <mingo@elte.hu>
Date:   Wed Apr 27 04:06:33 2011 +0200

description
-----------
Currently we fail without printing any error message on "perf stat -e task-clock-msecs".

The reason is that the task-clock event is matched and the "-msecs" postfix is assumed
to be an event modifier - but is not recognized.

This patch changes the code to be more informative:

 $ perf stat -e task-clock-msecs true
 invalid event modifier: '-msecs'
 Run 'perf list' for a list of valid events and modifiers

And restructures the return value of parse_event_modifier() to allow
the printing of all variants of invalid event modifiers.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>