stat: Make all displayed event names parseable as well
Patchwork-id: 38973
O-Subject: [PATCH RHEL6.2 BZ725522 19/67] perf stat: Make all displayed event
	names parseable as well
Bugzilla: 725522
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 725522
https://bugzilla.redhat.com/show_bug.cgi?id=725522

upstream status
---------------
commit 749141d926faf23ef811686a8050e7cf13dc223f
Author: Ingo Molnar <mingo@elte.hu>
Date:   Wed Apr 27 04:24:57 2011 +0200

description
-----------
Right now we display this by default:

          0.202204 task-clock-msecs         #      0.282 CPUs
                 0 context-switches         #      0.000 M/sec
                 0 CPU-migrations           #      0.000 M/sec
                85 page-faults              #      0.420 M/sec

The task-clock-msecs event cannot actually be passed back as an
event name, the event name we recognize is 'task-clock'.

So change the output of the cpu-clock and task-clock events
to be idempotent.

( Units should be printed out in the right-side column, if needed. )

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>