tools: Fix ommitted mmap data update on remap
Patchwork-id: 38957
O-Subject: [PATCH RHEL6.2 BZ725522 50/67] perf tools: Fix ommitted mmap data
	update on remap
Bugzilla: 725522
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 725522
https://bugzilla.redhat.com/show_bug.cgi?id=725522

upstream status
---------------
commit 998bedc8c56c6869de457c845cbd328592e5e82e
Author: Frederic Weisbecker <fweisbec@gmail.com>
Date:   Mon May 23 13:06:28 2011 +0200

description
-----------
Commit eac9eacee16 "perf tools: Check we are able to read the event
size on mmap" brought a check to ensure we can read the size of the
event before dereferencing it, and do a remap otherwise to move the
buffer forward.

However that remap was ommitting all the necessary work to
update the new page offset, head, and to unmap previous pages,
etc...

To fix this, gather all the code that fetches the event in a
seperate helper which does all the necessary checks about the
header/event size and tells us anytime a remap is needed.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>