events: initialize fd array to -1 instead of 0
Patchwork-id: 38969
O-Subject: [PATCH RHEL6.2 BZ725522 57/67] perf events: initialize fd array to -1
	instead of 0
Bugzilla: 725522
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 725522
https://bugzilla.redhat.com/show_bug.cgi?id=725522

upstream status
---------------
commit 4af4c9550ccaaf0b53013ff730bc15068ffe6abc
Author: David Ahern <dsahern@gmail.com>
Date:   Fri May 27 09:58:34 2011 -0600

description
-----------
perf_evsel__alloc_fd allocates an array of file descriptors with the
memory initialized to 0. The array has dimensions for cpus and threads.

Later, __perf_evsel__open calls sys_perf_event_open for each cpu and thread
dimensions. If the open fails for any of the cpus or threads then the fd's
for this event are closed and the fd entry in the array is set to -1. Now,
if the first attempt fails for the event (e.g., the event is not supported)
the remaining dimensions (cpu > 0 and thread > 0) are not touched and left
at the initialized value of 0.

builtin-stat catches ENOENT and ENOSYS failures and allows the command to
continue. The end result is that stat attempts to read from an fd of 0 which
of course is stdin and so the command hangs until you type ctrl-D.

Resolve by initializing the array to -1 since an fd < 0 is already
handled.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>