Fix prior ABI breakage in net_device
Patchwork-id: 39138
O-Subject: [RHEL 6.2 PATCH] net: Fix prior ABI breakage in net_device (bz 725716)
Bugzilla: 725716
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>
RH-Acked-by: Thomas Graf <tgraf@redhat.com>

From: Neil Horman <nhorman@tuxdriver.com>

Previous patches enabling hw qos and receive packet steering added data to the
net_device structure.  I was under the impression this would not break kabi, but
as it turns out it does because of how netdev_priv() works.  So that data needs
to not be at the tail of the structure.  jiri's XPS patch series adds a
netdev_extended field to move this and other like data to a safe place.  This
patch applies on top of those to unbreak kabi

Brew:
https://brewweb.devel.redhat.com/taskinfo?taskID=3517884

Testing:
Verified that nothing melts down after this change on amd-dinar.  Enabled rps on
the interface and pounded a bit of traffic.

ABI status:
Several symbols were rescued from the wild and are being nursed back to health.

Neil

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>