cifs: fix name parsing in CIFSSMBQAllEAs
Patchwork-id: 39145
O-Subject: [RHEL6 PATCH] BZ#725370: cifs: fix name parsing in CIFSSMBQAllEAs
Bugzilla: 725370
RH-Acked-by: Ian Kent <ikent@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: David Howells <dhowells@redhat.com>

The code that matches EA names in CIFSSMBQAllEAs is incorrect. It
uses strncmp to do the comparison with the length limited to the
name_len sent in the response.

Problem: Suppose we're looking for an attribute named "foobar" and
have an attribute before it in the EA list named "foo". The
comparison will succeed since we're only looking at the first 3
characters. Fix this by also comparing the length of the provided
ea_name with the name_len in the response. If they're not equal then
it shouldn't match.

Reported-by: Jian Li <jiali@redhat.com>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Reviewed-by: Pavel Shilovsky <piastryyy@gmail.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>