SELinux: return -ECONNREFUSED from ip_postroute to signal fatal error
Patchwork-id: 39151
O-Subject: [RHEL6 PATCH 3/5] SELinux: return -ECONNREFUSED from ip_postroute to
	signal fatal error
Bugzilla: 656952
RH-Acked-by: James Morris <jmorris@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>

RHBZ: 656952
upstream commit 1f1aaf82825865a50cef0b4722607abb12aeee52

The SELinux netfilter hooks just return NF_DROP if they drop a packet.  We
want to signal that a drop in this hook is a permanant fatal error and is not
transient.  If we do this the error will be passed back up the stack in some
places and applications will get a faster interaction that something went
wrong.

Signed-off-by: Eric Paris <eparis@redhat.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>