Silence warnings when compiling our git kernel on Fedora
Patchwork-id: 39158
O-Subject: [Patch RHEL6.2] Silence warnings when compiling our git kernel on
	Fedora
Bugzilla:
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Not strictly a necessary patch, but works and makes compiling our git
tree on a Fedora 15 or later host not *nearly* as painful as it used to
be.  It checks if you are running a newer gcc, and then disables the
warnings that the upstream kernel people have decided are just random
noise for the kernel.  Tested on my F15 workstation and makes a kernel
build *much* nicer.

No bug of course, just something for internal usability of our git tree.
  Given the way we do wholesale updates to part of our kernel, it seems
this isn't too unnecessary of a change to at least consider ;-)
commit 646ca297a1dc145e5b0fa838cc32d69fd6834984
Author: Doug Ledford <dledford@redhat.com>
Date:   Thu Jul 28 17:13:12 2011 -0400

    Backport changes for gcc-4.4 warning management

    Later gcc versions (such as in Fedora 15) will compile our internal
    kernel, but at the cost of a *huge* amount of noise from warning
    options that don't exist in the rhel6 compiler.  Upstream added these
    changes in order to silence newer gcc versions while still being able
    to see through the noise.  Add it to our tree too so we can run test
    compiles on non-rhel6 machines.

    Signed-off-by: Doug Ledford <dledford@redhat.com>

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>