bonding: fix string comparison errors
Patchwork-id: 39286
O-Subject: [PATCH RHEL6.2] bonding: fix string comparison errors
Bugzilla: 682946
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Thomas Graf <tgraf@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>

Customer reported that an error in the string comparison routines when
storing the primary and active slave would not detect a difference when
the current value was a subset of the new value.  This would mean that
someone trying to set the primary to 'eth10' would not be able to do it
if 'eth1' was the current value.

This is a backport of the following upstream commit:

f4bb2e9c4fa9e5fdddf90589703613fd1a9c519f bonding: fix string comparison errors

This will resolve RHBZ 682946.

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>