bonding: don't lock when copying/clearing VLAN list on slave
Patchwork-id: 39288
O-Subject: [PATCH RHEL6.2] bonding: don't lock when copying/clearing VLAN list on slave
Bugzilla: 722901
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Thomas Graf <tgraf@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>

Broadcom reported that holding the bond->lock while modifying vlan list
for a slave was causing problems for any driver's ndo_vlan* functions
that might sleep.  It turned out this lock was not necessary, so it was
dropped.

This is a backport of the following upstream commit:

03dc2f4c525afb9488edb687c2e1f7057d59b40e bonding: don't lock when copying/clearing VLAN list on slave

This will resolve RHBZ 722901.

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>