isci: merge ssp task management substates into primary state machine
Patchwork-id: 39396
O-Subject: [RHEL6.2 PATCH 92/179 v3] BZ 695954 isci: merge ssp task management
	substates into primary state machine
Bugzilla: 695954
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Dan Williams <dan.j.williams@intel.com>

Remove usage of the request substate machine for ssp task management
requests identified by:
	ireq->ttype == tmf_task && dev->dev_type == SAS_END_DEV;

The only routine that checks the base 'started' state is
scic_sds_io_request_tc_completion which calls the substate machine
handler if we are not in the 'started' state or we are 'started' and no
substate machine is defined.  This routine requires no conversion
because we have transitioned out of 'started' and the substate routine
will be called naturally as a result.

There are also no side effects of this conversion on exiting the
'started', state because it only stops the substate machine, which is no
longer relevant for this transaction type.

Reported-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>

BZ 695954

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>