isci: removing the kmalloc in smp request construct
Patchwork-id: 39405
O-Subject: [RHEL6.2 PATCH 131/179 v3] BZ 695954 isci: removing the kmalloc in
	smp request construct
Bugzilla: 695954
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Dave Jiang <dave.jiang@intel.com>

It doesn't look like there is any reason to do a kmalloc. We can do the
byte swap in place and avoid the allocation. This allow us to remove
a kmalloc and a memcpy.

Signed-off-by: Dave Jiang <dave.jiang@intel.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>

BZ 695954

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>