isci: kill isci_remote_device_change_state()
Patchwork-id: 39469
O-Subject: [RHEL6.2 PATCH 152/179 v3] BZ 695954 isci: kill
	isci_remote_device_change_state()
Bugzilla: 695954
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Dan Williams <dan.j.williams@intel.com>

Now that "stopping/stopped" are one in the same and signalled by a NULL device
pointer the rest of the device status infrastructure can be removed (->status
and ->state_lock).  The "not ready for i/o state" is replaced with a state
flag, and is evaluated under scic_lock so that we don't see transients from
taking the device reference to submitting the i/o.

This also fixes a potential leakage of can_queue slots in the rare case that
SAS_TASK_ABORTED is set at submission.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>

BZ 695954

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>