isci: Terminate dev requests on FIS err bit rx in NCQ
Patchwork-id: 39451
O-Subject: [RHEL6.2 PATCH 160/179 v3] BZ 695954 isci: Terminate dev requests on
	FIS err bit rx in NCQ
Bugzilla: 695954
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Jeff Skirvin <jeffrey.d.skirvin@intel.com>

When the remote device transitions to a not-ready state because of
an NCQ error condition, all outstanding requests to that device
are terminated and completed to libsas on the normal path.  The
device then waits for a READ LOG EXT command to issue on the task
management path.

Signed-off-by: Jeff Skirvin <jeffrey.d.skirvin@intel.com>
Signed-off-by: Dan Williams <dan.j.williams@intel.com>

BZ 695954

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>