mpt2sas: lockless mode
Patchwork-id: 39502
O-Subject: [RHEL6.2 PATCH] mpt2sas: lockless mode
Bugzilla: 717090
RH-Acked-by: Rob Evers <revers@redhat.com>

This is for bz#717090

This patch adds the option to the mpt2sas driver to work lock-less
A patch for this is posted upstream but differs from this one because
of KABI reasons. A partners testing shows significant performance improvement,
but LSI is still having some concerns. The patch below adds an option,
so the interested patners can test it, the lock-less mode isn't on by default
and a warning is added - I'm not sure though, if this is the right way.
Patch comes from LSI, I modified it for the KABI.

Tomas

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>