mremap: resurrect overflow check in move_page_tables()
Patchwork-id: 39513
O-Subject: [RHEL6.2 PATCH] mremap: resurrect overflow check in move_page_tables()
Bugzilla: 725926
RH-Acked-by: Johannes Weiner <jweiner@redhat.com>

I discussed this upstream with Johannes and Mel. Johannes was the
first to point out it may be there to avoid an overflow
condition. That condition can only happen if TASK_SIZE ends less than
PMD_SIZE from the overflow point. Sure it can't happen on x86 and
sparc also seems safe.

But now that we have a explanation for why that might have been
beneficial it's safer to resurrect it. I think the current code is
more correct for x86, but it's not worth taking risk for other archs
and these old_* addresses should be always page aligned.

For upstream I plan to modify it to still remove the -1 off by one
error and do a proper "delta" (not absolute like above) check that
will be safe even if next overflows (and removing the -1 that is
technically an off by one error that relays on stuff to be aligned to
PAGE_SIZE > 1). But I don't want to modify the code in RHEL6 so
returning to the previous code that did -1 sounds the safest.

On the plus side Mel reviewed the whole mremap optimization (Johannes
also reviewed a previous version) and didn't find any problem in the
THP bits.

Thanks,
Andrea

https://bugzilla.redhat.com/show_bug.cgi?id=725926

===
Subject: mremap: resurrect overflow check in move_page_tables()

From: Andrea Arcangeli <aarcange@redhat.com>

If any arch sets the TASK_SIZE at less than PMD_SIZE from the end of
the userland address space, "next" could overflow and that -1 could have
made a difference. Not on x86 and sparc that can use all 32bit of
userland address space (because the kernel address space is separate)
also don't seem to be in trouble.

Now that Johannes had a possible explanation for the -1 it's safer to
resurrect it. It's a noop for x86 32/64bit.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>