vxge: allow rebinding the driver with a different number of SR-IOV VFs
Patchwork-id: 39536
O-Subject: [RHEL6.2 PATCH] vxge: allow rebinding the driver with a different
	number of SR-IOV VFs
Bugzilla: 694742
RH-Acked-by: Andy Gospodarek <gospo@redhat.com>
RH-Acked-by: Michael S. Tsirkin <mst@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>

On Fri, 5 Aug 2011 16:03:51 -0400 Andy Gospodarek wrote:
> It seems like sriov should actually be disabled and
> re-enabled when the driver is loaded if the number of vfs changes.

OK. Let's add this on top of the previous patch.

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=694742

When the requested number of VFs is different from the current state,
disable and re-enable SR-IOV with the new number.

Don't touch SR-IOV if the right number of VFs is already enabled.

No need to disable SR-IOV in the error path.

Suggested by Michael S. Tsirkin and Andy Gospodarek.

I will submit upstream if this looks OK.

Brew build in progress:
https://brewweb.devel.redhat.com/taskinfo?taskID=3541848

Tested by me. I verified that the number of VFs can change when
reloading the driver (to test it easily I added a module parameter to
override num_vfs, not included in this patch).
I did not test with virtual guests, but I'm sure they won't be happy if
their assigned VFs disappear. This will happen only when the driver
is rebound to the PF _and_ the function mode (which determines num_vfs)
of the card changes, so it's still an improvement over what we had
before.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>