nfs: don't use d_move in nfs_async_rename_done
Patchwork-id: 39559
O-Subject: [RHEL6.2 PATCH] BZ#729434: nfs: don't use d_move in nfs_async_rename_done
Bugzilla: 729434
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

This patch fixes a potential issue that Al spotted upstream with the
async sillyrename patches. Since those were backported for 6.1,
it's also a potential problem there too.

If the task that initiated the sillyrename ends up being killed by a
fatal signal, then it will eventually return back to userspace and end
up releasing the i_mutex. d_move however needs to be done while holding
the i_mutex.

Instead of using d_move here, just unhash the old and new dentries to
prevent them from being found by lookups. With this change though, the
dentries are now incorrect post-rename and do not reflect the actual
name of the file on the server. I'm proceeding under the assumption
that since they are unhashed that this isn't really a problem.

In order for the sillydelete to still work though, the dname must be
copied earlier when setting up the sillydelete info, and the name must
be recopied if the sillydelete info has to be moved to a new dentry.

Reported-by: Al Viro <viro@ZenIV.linux.org.uk>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Upstream commit: 73ca1001ed6881b476e8252adcd0eede1ea368ea

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>